key will be deleted after 2nd time

+++ verification/views.py
@@ -1,12 +1,28 @@
 # verification/views.py
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 from django.http import JsonResponse
 from .models import SerialKey
+
+
+def check_and_update_serial_key(serial_key):
+    key_object = get_object_or_404(SerialKey, key=serial_key, used_time__lt=2)
+
+    # Increment used_time
+    key_object.used_time += 1
+
+    # If used_time is 2 or more, delete the key
+    if key_object.used_time >= 2:
+        key_object.delete()
+    else:
+        key_object.save()
+
+    return key_object.used_time
+
 
 def check_serial_key_in_database(serial_key):
     try:
         key_object = SerialKey.objects.get(key=serial_key)
-        key_object.delete()
+       # key_object.delete()
         return True
     except SerialKey.DoesNotExist:
         return False
@@ -19,6 +35,7 @@ def verify_authenticity(request):
         is_authentic = check_serial_key_in_database(serial_key)
 
         if is_authentic:
+            used_time = check_and_update_serial_key(serial_key)
             response_data = {'result': 'authentic'}
         else:
             response_data = {'result': 'not_authentic'}
